+++ src/data/id/handle/mod.rs
@@ -3,8 +3,8 @@
 //!
 //
 
-// mod define; // WIP TEMP handle2!
-mod legacy; // OLD handle! (Span)
+// mod define; // WIP handle2!
+mod span; // handle_span!
 
 #[cfg(any(test, feature = "_docs_examples"))]
 mod _example; // HandleSpanExample
@@ -14,7 +14,7 @@ crate::structural_mods! { // _mods
         #[doc(inline)]
         pub use super::{
             // define::*,
-            legacy::*,
+            span::*,
         };
         #[cfg(any(test, feature = "_docs_examples"))]
         pub use super::_example::*;