+++ src/data/layout/mod.rs
@@ -8,7 +8,6 @@
 //
 
 mod collection; // DataCollection
-// mod pool; // WIP
 mod queue;
 mod sort; // Sort
 mod stack;
@@ -28,7 +27,6 @@ crate::structural_mods! { // _mods, _pub_mods, _crate_internals
     _mods {
         pub use super::{
             collection::*,
-            // pool::_all::*,
             queue::_all::*,
             sort::_all::*,
             stack::_all::*,