+++ src/data/store/arena/typed/mod.rs
@@ -1,0 +1,14 @@
+// devela/src/data/store/arena/typed/mod.rs
+//
+//!
+//
+
+// mod define; // arena_typed!
+
+crate::structural_mods! { // _mods
+    _mods {
+        // pub use super::{
+        //     define::*,
+        // };
+    }
+}