+++ src/data/store/mod.rs
@@ -7,14 +7,29 @@
 #![doc = crate::_QUO_DATA_ID!()]
 //
 
-// pub mod cache; // Stores that retain values to reduce recomputation or retrieval cost
-// pub mod db; // Persistent queryable stores with schema and transactional semantics
+mod arena; // WIP
+// mod cache; // Stores that retain values to reduce recomputation or retrieval cost
+// mod db; // Persistent queryable stores with schema and transactional semantics
 pub mod key; // Keyed stores and lookup structures
+// mod pool; // WIP
 
-crate::structural_mods! { // __pub_mods
+crate::structural_mods! { // _mods, _pub_mods, _hidden
+    _mods {
+        pub use super::{
+            arena::_all::*,
+            // cache::_all::*,
+            // db::_all::*,
+            // pool::_all::*,
+        };
+    }
     _pub_mods {
         pub use super::{
             key::_all::*,
+        };
+    }
+    _hidden {
+        pub use super::{
+            arena::_hidden::*,
         };
     }
 }