+++ src/data/store/pool/mod.rs
@@ -1,0 +1,11 @@
+// devela/src/data/layout/pool/mod.rs
+//
+//!
+//
+
+crate::structural_mods! { // _mods
+    _mods {
+        // pub use super::{
+        // };
+    }
+}